+++ frontend/src/routes.tsx
@@ -10,6 +10,7 @@ import { DipoleHeightLabPage } from "./pages/DipoleHeightLabPage";
 import { AntennaTemplateStudioPage } from "./pages/AntennaTemplateStudioPage";
 import { VerticalAntennasPage } from "./pages/VerticalAntennasPage";
 import { YagiBeamModelsPage } from "./pages/YagiBeamModelsPage";
+import { LoopAndHexbeamModelsPage } from "./pages/LoopAndHexbeamModelsPage";
 
 export function AppRoutes() {
   return (
@@ -21,6 +22,7 @@ export function AppRoutes() {
       <Route path="/antenna-templates" element={<AntennaTemplateStudioPage />} />
       <Route path="/vertical-antennas" element={<VerticalAntennasPage />} />
       <Route path="/yagi-beams" element={<YagiBeamModelsPage />} />
+      <Route path="/loop-and-hexbeam-models" element={<LoopAndHexbeamModelsPage />} />
       <Route path="/library" element={<LibraryPage />} />
       <Route path="/learn" element={<LearnPage />} />
       <Route path="/about" element={<AboutPage />} />